Initial Linux sound support
* Uses ALSA/libasound but we may have to use something else as the Linux
  sound subsystem SUUUUUUCKS!!!!! Talk about a complete disappointment:
  10 years with the promise of something "oh so much better" than OSS and
  ALSA can't even do PCM multichannel mixing OR PCM volume control? WTF?!?

+++ graphics.c
@@ -1584,7 +1584,7 @@ void display_fps(uint64_t frames_duration, uint64_t nb_frames)
         li = nb_frames;
     }
 
-    sprintf(s_fps, "%3lldFPS", li*1000/lf);
+    sprintf(s_fps, "%3lldFPS", (unsigned long long) li*1000LL/lf);
 
     glColor3f(0.3f, 0.4f, 1.0f);
